client/requirements.txt
flask
python-dotenv
pyjwt==1.7.1
requests==2.22.0